+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
 
 * Make categories non required
 * Fix tests with parler>=1.6
+* Use all_languages_column to admin
 
 0.6.3 (2015-12-22)
 ++++++++++++++++++